refactor: changing package on repository's models
