+++ src/main/java/it/gov/pagopa/fdr/repository/FdrFlowRepository.java
@@ -5,9 +5,9 @@ import io.quarkus.panache.common.Page;
 import io.quarkus.panache.common.Parameters;
 import io.quarkus.panache.common.Sort;
 import io.quarkus.panache.common.Sort.Direction;
-import it.gov.pagopa.fdr.repository.entity.common.Repository;
-import it.gov.pagopa.fdr.repository.entity.common.RepositoryPagedResult;
-import it.gov.pagopa.fdr.repository.entity.common.SortField;
+import it.gov.pagopa.fdr.repository.common.Repository;
+import it.gov.pagopa.fdr.repository.common.RepositoryPagedResult;
+import it.gov.pagopa.fdr.repository.common.SortField;
 import it.gov.pagopa.fdr.repository.entity.flow.FdrFlowEntity;
 import it.gov.pagopa.fdr.repository.entity.flow.projection.FdrFlowIdProjection;
 import it.gov.pagopa.fdr.repository.enums.FlowStatusEnum;
