+++ src/main/java/it/gov/pagopa/fdr/repository/FdrPaymentRepository.java
@@ -7,9 +7,9 @@ import io.quarkus.panache.common.Page;
 import io.quarkus.panache.common.Parameters;
 import io.quarkus.panache.common.Sort;
 import io.quarkus.panache.common.Sort.Direction;
-import it.gov.pagopa.fdr.repository.entity.common.Repository;
-import it.gov.pagopa.fdr.repository.entity.common.RepositoryPagedResult;
-import it.gov.pagopa.fdr.repository.entity.common.SortField;
+import it.gov.pagopa.fdr.repository.common.Repository;
+import it.gov.pagopa.fdr.repository.common.RepositoryPagedResult;
+import it.gov.pagopa.fdr.repository.common.SortField;
 import it.gov.pagopa.fdr.repository.entity.payment.FdrPaymentEntity;
 import it.gov.pagopa.fdr.repository.exception.PersistenceFailureException;
 import it.gov.pagopa.fdr.repository.exception.TransactionRollbackException;
