+++ src/main/java/it/gov/pagopa/fdr/repository/common/SortField.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.repository.entity.common;
+package it.gov.pagopa.fdr.repository.common;
 
 import io.quarkus.panache.common.Sort.Direction;
 import lombok.Data;
